move src/model to models, delete model writeups, add __init__.py's throughout
