+++ models/segment_counting/convnet1/.gitignore
@@ -1,0 +1,2 @@
+.ipynb_checkpoints
+saves
