+++ models/segmentation/convnet3/.gitignore
@@ -1,0 +1,2 @@
+.ipynb_checkpoints
+saves
