--- src/model/segment_counting/convnet1/.gitignore
@@ -1,2 +1,0 @@
-.ipynb_checkpoints
-saves
