--- src/model/segmentation/convnet2/.gitignore
@@ -1,2 +1,0 @@
-.ipynb_checkpoints
-saves
